Refactor: Move out the async thunks to their respective files
