+++ src/services/gameCalls.js
@@ -1,0 +1,13 @@
+/* eslint-disable consistent-return */
+/* eslint-disable import/prefer-default-export */
+import axios from 'axios';
+import { createAsyncThunk } from '@reduxjs/toolkit';
+
+export const fetchGames = createAsyncThunk('games/fetchGames', async () => {
+  try {
+    const response = await axios.get('http://127.0.0.1:3000/games');
+    return response.data;
+  } catch (error) {
+    console.log(error);
+  }
+});
